Implement a queue to sync session data
